Bug #1112572 update vdsp faceid app
[root cause  ] new request
[changes     ] update
[side effects]N\A
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:sharkl5pro
[test case]:faceid
[reviewers   ] kevin.yang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ vdsp/Cadence/xrp/xrp-host/inc/xrp_api.h
@@ -104,7 +104,7 @@ typedef struct
 	int x, y, width, height;
 	int yawAngle, pitchAngle;
 	int ret;
-	unsigned long facepoint_addr;
+	unsigned int facepoint_addr;
 }FACEID_INFO;
 #define XRP_NAMESPACE_ID_SIZE	32
